Fix return type of for get-members route
the returned type was not the correct PHP object (but I did not
pay attention to it because PHP does not make it possible to
declare template type.

Also changed the list of required fields in the response because
it was inconsistent to have a field optional in the db schema
but mandatory in the OAS response.