change namespaces
the previous namespace / layout was suggested by openApiGenerator
generated README, but it proved to not always work fine with Symfony
cache: sometimes after a 1st query, all the subsequent one failed
because Symfony complained about not being able to locate our classes.

According to the error messages it got the namespace wrong on those
subsequent queries, which hinted that the namespace / layout was a
part of the problem.
It turns out that, indeed, changing it seems to fix the issue for good.

Note that in the process I'm also committing several changes that I had
locally that represented unfinished features. It's not great, but
at least everything should work (or at least not throw exception even
if some endpoint don't do all the work they will eventually do)