Can access directly any front url
Before this commit Symfony intercepts queries that do not match
a resource in the 'public' directory and tries to find an endpoint
that it knows.
So when we query eg "/account", since it's a route known by angular
but not by Symfony, then Symfony replies with a 404.

With this commit, we intercept queries that do not match a resource
in the public directory, and we serve index.html when we detect the query
is not meant to be handled by Symfony (that way we serve the angular app,
and its router subsequently display the expected page)